requirements.txt
python-dotenv >= 1.0.1            # Minimum version 1.0.1